+++ Tests/SentryTests/SentryNSURLRequestTests.swift
@@ -1,4 +1,4 @@
-@testable import Sentry
+@_spi(Private) @testable import Sentry
 import SentryTestUtils
 import XCTest
 
@@ -10,13 +10,13 @@ class SentryNSURLRequestTests: XCTestCase {
     }
     
     func testRequestWithEnvelopeEndpoint() throws {
-        let request = try SentryNSURLRequest(envelopeRequestWith: SentryNSURLRequestTests.dsn(), andData: Data())
+        let request = try SentryURLRequestFactory.envelopeRequest(with: SentryNSURLRequestTests.dsn(), data: Data())
         let string = try XCTUnwrap(request.url?.absoluteString as? NSString)
         XCTAssert(string.hasSuffix("/envelope/"))
     }
     
     func testRequestWithEnvelopeEndpoint_hasUserAgentWithSdkNameAndVersion() {
-        let request = try! SentryNSURLRequest(envelopeRequestWith: SentryNSURLRequestTests.dsn(), andData: Data())
+        let request = try! SentryURLRequestFactory.envelopeRequest(with: SentryNSURLRequestTests.dsn(), data: Data())
         XCTAssertEqual(request.allHTTPHeaderFields?["User-Agent"], "\(SentryMeta.sdkName)/\(SentryMeta.versionString)")
     }
 }