+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -156,8 +156,6 @@
 #import "SentryNSNotificationCenterWrapper.h"
 #import "SentryNSProcessInfoWrapper.h"
 #import "SentryNSTimerFactory.h"
-#import "SentryNSURLRequest.h"
-#import "SentryNSURLRequestBuilder.h"
 #import "SentryNSURLSessionTaskSearch.h"
 #import "SentryNetworkTracker.h"
 #import "SentryNetworkTrackingIntegration.h"